+++ pyrcn/extreme_learning_machine/__init__.py
@@ -9,10 +9,8 @@ The :mod:`pyrcn.elm` module includes Extreme Learning Machine algorithms.
 #         the pyrcn community. ELM are copyright of their respective owners.
 # License: BSD 3-Clause (C) TU Dresden 2020
 
-from pyrcn.extreme_learning_machine._base import InputToNode
 from pyrcn.extreme_learning_machine._elm import ELMClassifier, ELMRegressor
 
 __all__ = ['ELMClassifier',
-           'ELMRegressor',
-           'InputToNode'
+           'ELMRegressor'
            ]
